Align with PR feedback( Peng)

+++ src/Microsoft.Health.Dicom.SqlServer/Features/Schema/Migrations/2.diff.sql
@@ -129,7 +129,7 @@ CREATE UNIQUE CLUSTERED INDEX IXC_CustomTagDateTime ON dbo.CustomTagDateTime
           For example, with multiple instances in a series, while indexing a series level tag,
           the Watermark is used to ensure that if there are different values between instances,
           the value on the instance with the highest watermark wins.
-	Note: The primary key is designed on the assumption that tags only occur once in an instance.
+    Note: The primary key is designed on the assumption that tags only occur once in an instance.
 **************************************************************/
 CREATE TABLE dbo.CustomTagPersonName (
     TagKey                  INT                  NOT NULL, --FK
@@ -176,7 +176,7 @@ GO
 *************************************************************/
 CREATE TYPE dbo.InsertStringCustomTagTableType_1 AS TABLE
 (
-    TagKey                     BIGINT,
+    TagKey                     INT,
     TagValue                   NVARCHAR(64),
     TagLevel                   TINYINT
 )
@@ -187,7 +187,7 @@ GO
 *************************************************************/
 CREATE TYPE dbo.InsertDoubleCustomTagTableType_1 AS TABLE
 (
-    TagKey                     BIGINT,
+    TagKey                     INT,
     TagValue                   FLOAT(53),
     TagLevel                   TINYINT
 )
@@ -198,8 +198,8 @@ GO
 *************************************************************/
 CREATE TYPE dbo.InsertBigIntCustomTagTableType_1 AS TABLE
 (
-    TagKey                     BIGINT,
-    TagValue                BIGINT,
+    TagKey                     INT,
+    TagValue                   BIGINT,
     TagLevel                   TINYINT
 )
 GO
@@ -209,8 +209,8 @@ GO
 *************************************************************/
 CREATE TYPE dbo.InsertDateTimeCustomTagTableType_1 AS TABLE
 (
-    TagKey                     BIGINT,
-    TagValue              DATETIME2(7),
+    TagKey                     INT,
+    TagValue                   DATETIME2(7),
     TagLevel                   TINYINT
 )
 GO
@@ -220,8 +220,8 @@ GO
 *************************************************************/
 CREATE TYPE dbo.InsertPersonNameCustomTagTableType_1 AS TABLE
 (
-    TagKey                     BIGINT,
-    TagValue            NVARCHAR(200)        COLLATE SQL_Latin1_General_CP1_CI_AI,
+    TagKey                     INT,
+    TagValue                   NVARCHAR(200)        COLLATE SQL_Latin1_General_CP1_CI_AI,
     TagLevel                   TINYINT
 )
 GO
@@ -230,7 +230,7 @@ GO
     Sequence for generating sequential unique ids
 **************************************************************/
 CREATE SEQUENCE dbo.TagKeySequence
-    AS BIGINT
+    AS INT
     START WITH 1
     INCREMENT BY 1
     MINVALUE 1
@@ -264,7 +264,7 @@ AS
         FROM dbo.CustomTag WITH(HOLDLOCK) 
         INNER JOIN @customTags input 
         ON input.TagPath = dbo.CustomTag.TagPath 
-	    
+        
         IF @@ROWCOUNT <> 0
             THROW 50409, 'custom tag(s) already exist', 1 
 
@@ -500,178 +500,178 @@ AS
 
     -- Insert Custom Tags
     DECLARE @rowCount AS BIGINT
+    -- This variable is used for not returning value from SELECT.
     DECLARE @temp AS INT
 
-    -- String tags
+    -- String Key tags
     SET @rowCount = (SELECT COUNT(TagKey) FROM @stringCustomTags)
-	IF @rowCount <> 0
-	BEGIN
+    IF @rowCount <> 0
+    BEGIN
         -- Validate and lock custom tags
-	    SELECT @temp = dbo.CustomTag.TagKey
+        SELECT @temp = dbo.CustomTag.TagKey
         FROM dbo.CustomTag WITH (UPDLOCK)
-	    INNER JOIN @stringCustomTags input
-	    ON input.TagKey = dbo.CustomTag.TagKey 
-	    AND ( dbo.CustomTag.TagStatus = 0 OR  dbo.CustomTag.TagStatus = 1)
-	
-	    IF @@ROWCOUNT<>@rowCOUNT
-	        THROW 50409, 'Custom Tags have been changed', 10
-	
-	    -- Merge into dbo.CustomTagString table        
-	    MERGE INTO dbo.CustomTagString AS T
-	    USING @stringCustomTags AS S
-	    ON T.TagKey=S.TagKey
-	    AND T.StudyKey=@studyKey AND
-	    ISNULL(T.SeriesKey,@seriesKey) = @seriesKey AND
-	    ISNULL(T.InstanceKey,@instanceKey) = @instanceKey
+        INNER JOIN @stringCustomTags input
+        ON input.TagKey = dbo.CustomTag.TagKey 
+        AND ( dbo.CustomTag.TagStatus = 0 OR  dbo.CustomTag.TagStatus = 1)
+    
+        IF @@ROWCOUNT<>@rowCOUNT
+            THROW 50409, 'Custom Tags have been changed', 10
+    
+        -- Merge into dbo.CustomTagString table        
+        MERGE INTO dbo.CustomTagString AS T
+        USING @stringCustomTags AS S
+        ON T.TagKey = S.TagKey
+        AND T.StudyKey = @studyKey
+        AND ISNULL(T.SeriesKey, @seriesKey) = @seriesKey 
+        AND ISNULL(T.InstanceKey, @instanceKey) = @instanceKey
         WHEN MATCHED THEN 
             UPDATE SET T.Watermark = @newWatermark, T.TagValue = S.TagValue
-	    WHEN NOT MATCHED THEN 
+        WHEN NOT MATCHED THEN 
             INSERT (TagKey, TagValue, StudyKey, SeriesKey, InstanceKey, Watermark)
-	        VALUES(
-	        S.TagKey,
-	        S.TagValue,
-	        @studyKey,
-	        (CASE WHEN S.TagLevel<>2 THEN @seriesKey ELSE NULL END),
-	        (CASE WHEN S.TagLevel=0 THEN @instanceKey ELSE NULL END),
-	        @newWatermark);        
-	END
+            VALUES(
+            S.TagKey,
+            S.TagValue,
+            @studyKey,
+            (CASE WHEN S.TagLevel<>2 THEN @seriesKey ELSE NULL END),
+            (CASE WHEN S.TagLevel=0 THEN @instanceKey ELSE NULL END),
+            @newWatermark);        
+    END
 
-    -- BigInt tags
+    -- BigInt Key tags
     SET @rowCount = (SELECT COUNT(TagKey) FROM @bigintCustomTags)
-	IF @rowCount <> 0
-	BEGIN
+    IF @rowCount <> 0
+    BEGIN
         -- Validate and lock custom tags
-	    SELECT @temp = dbo.CustomTag.TagKey
+        SELECT @temp = dbo.CustomTag.TagKey
         FROM dbo.CustomTag WITH (UPDLOCK)
-	    INNER JOIN @bigintCustomTags input
-	    ON input.TagKey = dbo.CustomTag.TagKey 
-	    AND ( dbo.CustomTag.TagStatus = 0 OR  dbo.CustomTag.TagStatus = 1)
-	
-	    IF @@ROWCOUNT<>@rowCOUNT
-	        THROW 50409, 'Custom Tags have been changed', 10
-	
-	    -- Merge into dbo.CustomTagBigInt table        
-	    MERGE INTO dbo.CustomTagBigInt AS T
-	    USING @bigintCustomTags AS S
-	    ON T.TagKey=S.TagKey
-	    AND T.StudyKey=@studyKey AND
-	    ISNULL(T.SeriesKey,@seriesKey) = @seriesKey AND
-	    ISNULL(T.InstanceKey,@instanceKey) = @instanceKey
+        INNER JOIN @bigintCustomTags input
+        ON input.TagKey = dbo.CustomTag.TagKey 
+        AND ( dbo.CustomTag.TagStatus = 0 OR  dbo.CustomTag.TagStatus = 1)
+    
+        IF @@ROWCOUNT<>@rowCOUNT
+            THROW 50409, 'Custom Tags have been changed', 10
+    
+        -- Merge into dbo.CustomTagBigInt table        
+        MERGE INTO dbo.CustomTagBigInt AS T
+        USING @bigintCustomTags AS S
+        ON T.TagKey = S.TagKey
+        AND T.StudyKey = @studyKey
+        AND ISNULL(T.SeriesKey, @seriesKey) = @seriesKey 
+        AND ISNULL(T.InstanceKey, @instanceKey) = @instanceKey
         WHEN MATCHED THEN 
             UPDATE SET T.Watermark = @newWatermark, T.TagValue = S.TagValue
-	    WHEN NOT MATCHED THEN 
+        WHEN NOT MATCHED THEN 
             INSERT (TagKey, TagValue, StudyKey, SeriesKey, InstanceKey, Watermark)
-	        VALUES(
-	        S.TagKey,
-	        S.TagValue,
-	        @studyKey,
-	        (CASE WHEN S.TagLevel<>2 THEN @seriesKey ELSE NULL END),
-	        (CASE WHEN S.TagLevel=0 THEN @instanceKey ELSE NULL END),
-	        @newWatermark);        
-	END
+            VALUES(
+            S.TagKey,
+            S.TagValue,
+            @studyKey,
+            (CASE WHEN S.TagLevel<>2 THEN @seriesKey ELSE NULL END),
+            (CASE WHEN S.TagLevel=0 THEN @instanceKey ELSE NULL END),
+            @newWatermark);        
+    END
 
-    -- Double tags
+    -- Double Key tags
     SET @rowCount = (SELECT COUNT(TagKey) FROM @doubleCustomTags)
-	IF @rowCount <> 0
-	BEGIN
+    IF @rowCount <> 0
+    BEGIN
         -- Validate and lock custom tags
-	    SELECT @temp = dbo.CustomTag.TagKey
+        SELECT @temp = dbo.CustomTag.TagKey
         FROM dbo.CustomTag WITH (UPDLOCK)
-	    INNER JOIN @doubleCustomTags input
-	    ON input.TagKey = dbo.CustomTag.TagKey 
-	    AND ( dbo.CustomTag.TagStatus = 0 OR  dbo.CustomTag.TagStatus = 1)
-	
-	    IF @@ROWCOUNT<>@rowCOUNT
-	        THROW 50409, 'Custom Tags have been changed', 10
-	
-	    -- Merge into dbo.CustomTagDouble table        
-	    MERGE INTO dbo.CustomTagDouble AS T
-	    USING @doubleCustomTags AS S
-	    ON T.TagKey=S.TagKey
-	    AND T.StudyKey=@studyKey AND
-	    ISNULL(T.SeriesKey,@seriesKey) = @seriesKey AND
-	    ISNULL(T.InstanceKey,@instanceKey) = @instanceKey
+        INNER JOIN @doubleCustomTags input
+        ON input.TagKey = dbo.CustomTag.TagKey 
+        AND ( dbo.CustomTag.TagStatus = 0 OR  dbo.CustomTag.TagStatus = 1)
+    
+        IF @@ROWCOUNT<>@rowCOUNT
+            THROW 50409, 'Custom Tags have been changed', 10
+    
+        -- Merge into dbo.CustomTagDouble table        
+        MERGE INTO dbo.CustomTagDouble AS T
+        USING @doubleCustomTags AS S
+        ON T.TagKey = S.TagKey
+        AND T.StudyKey = @studyKey
+        AND ISNULL(T.SeriesKey, @seriesKey) = @seriesKey 
+        AND ISNULL(T.InstanceKey, @instanceKey) = @instanceKey
         WHEN MATCHED THEN 
             UPDATE SET T.Watermark = @newWatermark, T.TagValue = S.TagValue
-	    WHEN NOT MATCHED THEN 
+        WHEN NOT MATCHED THEN 
             INSERT (TagKey, TagValue, StudyKey, SeriesKey, InstanceKey, Watermark)
-	        VALUES(
-	        S.TagKey,
-	        S.TagValue,
-	        @studyKey,
-	        (CASE WHEN S.TagLevel<>2 THEN @seriesKey ELSE NULL END),
-	        (CASE WHEN S.TagLevel=0 THEN @instanceKey ELSE NULL END),
-	        @newWatermark);        
-	END
+            VALUES(
+            S.TagKey,
+            S.TagValue,
+            @studyKey,
+            (CASE WHEN S.TagLevel<>2 THEN @seriesKey ELSE NULL END),
+            (CASE WHEN S.TagLevel=0 THEN @instanceKey ELSE NULL END),
+            @newWatermark);        
+    END
 
-    -- DateTime tags
+    -- DateTime Key tags
     SET @rowCount = (SELECT COUNT(TagKey) FROM @dateTimeCustomTags)
-	IF @rowCount <> 0
-	BEGIN
+    IF @rowCount <> 0
+    BEGIN
         -- Validate and lock custom tags
-	    SELECT @temp = dbo.CustomTag.TagKey
+        SELECT @temp = dbo.CustomTag.TagKey
         FROM dbo.CustomTag WITH (UPDLOCK)
-	    INNER JOIN @dateTimeCustomTags input
-	    ON input.TagKey = dbo.CustomTag.TagKey 
-	    AND ( dbo.CustomTag.TagStatus = 0 OR  dbo.CustomTag.TagStatus = 1)
-	
-	    IF @@ROWCOUNT<>@rowCOUNT
-	        THROW 50409, 'Custom Tags have been changed', 10
-	
-	    -- Merge into dbo.CustomTagDateTime table        
-	    MERGE INTO dbo.CustomTagDateTime AS T
-	    USING @dateTimeCustomTags AS S
-	    ON T.TagKey=S.TagKey
-	    AND T.StudyKey=@studyKey AND
-	    ISNULL(T.SeriesKey,@seriesKey) = @seriesKey AND
-	    ISNULL(T.InstanceKey,@instanceKey) = @instanceKey
+        INNER JOIN @dateTimeCustomTags input
+        ON input.TagKey = dbo.CustomTag.TagKey 
+        AND ( dbo.CustomTag.TagStatus = 0 OR  dbo.CustomTag.TagStatus = 1)
+    
+        IF @@ROWCOUNT<>@rowCOUNT
+            THROW 50409, 'Custom Tags have been changed', 10
+    
+        -- Merge into dbo.CustomTagDateTime table        
+        MERGE INTO dbo.CustomTagDateTime AS T
+        USING @dateTimeCustomTags AS S
+        ON T.TagKey = S.TagKey
+        AND T.StudyKey = @studyKey
+        AND ISNULL(T.SeriesKey, @seriesKey) = @seriesKey 
+        AND ISNULL(T.InstanceKey, @instanceKey) = @instanceKey
         WHEN MATCHED THEN 
             UPDATE SET T.Watermark = @newWatermark, T.TagValue = S.TagValue
-	    WHEN NOT MATCHED THEN 
+        WHEN NOT MATCHED THEN 
             INSERT (TagKey, TagValue, StudyKey, SeriesKey, InstanceKey, Watermark)
-	        VALUES(
-	        S.TagKey,
-	        S.TagValue,
-	        @studyKey,
-	        (CASE WHEN S.TagLevel<>2 THEN @seriesKey ELSE NULL END),
-	        (CASE WHEN S.TagLevel=0 THEN @instanceKey ELSE NULL END),
-	        @newWatermark);        
-	END
+            VALUES(
+            S.TagKey,
+            S.TagValue,
+            @studyKey,
+            (CASE WHEN S.TagLevel<>2 THEN @seriesKey ELSE NULL END),
+            (CASE WHEN S.TagLevel=0 THEN @instanceKey ELSE NULL END),
+            @newWatermark);        
+    END
 
-    -- PersonName tags
+    -- PersonName Key tags
     SET @rowCount = (SELECT COUNT(TagKey) FROM @personNameCustomTags)
-	IF @rowCount <> 0
-	BEGIN
+    IF @rowCount <> 0
+    BEGIN
         -- Validate and lock custom tags
-	    SELECT @temp = dbo.CustomTag.TagKey
+        SELECT @temp = dbo.CustomTag.TagKey
         FROM dbo.CustomTag WITH (UPDLOCK)
-	    INNER JOIN @personNameCustomTags input
-	    ON input.TagKey = dbo.CustomTag.TagKey 
-	    AND ( dbo.CustomTag.TagStatus = 0 OR  dbo.CustomTag.TagStatus = 1)
-	
-	    IF @@ROWCOUNT<>@rowCOUNT
-	        THROW 50409, 'Custom Tags have been changed', 10
-	
-	    -- Merge into dbo.CustomTagPersonName table        
-	    MERGE INTO dbo.CustomTagPersonName AS T
-	    USING @personNameCustomTags AS S
-	    ON T.TagKey=S.TagKey
-	    AND T.StudyKey=@studyKey AND
-	    ISNULL(T.SeriesKey,@seriesKey) = @seriesKey AND
-	    ISNULL(T.InstanceKey,@instanceKey) = @instanceKey
+        INNER JOIN @personNameCustomTags input
+        ON input.TagKey = dbo.CustomTag.TagKey 
+        AND ( dbo.CustomTag.TagStatus = 0 OR  dbo.CustomTag.TagStatus = 1)
+    
+        IF @@ROWCOUNT<>@rowCOUNT
+            THROW 50409, 'Custom Tags have been changed', 10
+    
+        -- Merge into dbo.CustomTagPersonName table        
+        MERGE INTO dbo.CustomTagPersonName AS T
+        USING @personNameCustomTags AS S
+        ON T.TagKey = S.TagKey
+        AND T.StudyKey = @studyKey
+        AND ISNULL(T.SeriesKey, @seriesKey) = @seriesKey 
+        AND ISNULL(T.InstanceKey, @instanceKey) = @instanceKey
         WHEN MATCHED THEN 
             UPDATE SET T.Watermark = @newWatermark, T.TagValue = S.TagValue
-	    WHEN NOT MATCHED THEN 
+        WHEN NOT MATCHED THEN 
             INSERT (TagKey, TagValue, StudyKey, SeriesKey, InstanceKey, Watermark)
-	        VALUES(
-	        S.TagKey,
-	        S.TagValue,
-	        @studyKey,
-	        (CASE WHEN S.TagLevel<>2 THEN @seriesKey ELSE NULL END),
-	        (CASE WHEN S.TagLevel=0 THEN @instanceKey ELSE NULL END),
-	        @newWatermark);        
-	END
-
+            VALUES(
+            S.TagKey,
+            S.TagValue,
+            @studyKey,
+            (CASE WHEN S.TagLevel<>2 THEN @seriesKey ELSE NULL END),
+            (CASE WHEN S.TagLevel=0 THEN @instanceKey ELSE NULL END),
+            @newWatermark);        
+    END
 
     -- Insert Instance
     INSERT INTO dbo.Instance
